consolidate middleware tests
