+++ shared/test/middleware/files_api_test_base.rb
@@ -1,9 +1,9 @@
-require_relative 'test_helper' # Must be required first to establish load paths
+require_relative '../test_helper' # Must be required first to establish load paths
 require 'mocha/mini_test'
 require 'files_api'
 require 'channels_api'
 require 'cdo/aws/s3'
-require_relative 'spy_newrelic_agent'
+require_relative '../spy_newrelic_agent'
 
 #
 # Base class for tests against the FilesApi (which include SourcesTest,
