+++ shared/test/middleware/files_api_test_helper.rb
