+++ shared/test/middleware/helpers/test_library_bucket.rb
@@ -1,5 +1,5 @@
-require_relative '../../files_api_test_base' # Must be required first to establish load paths
-require_relative '../../files_api_test_helper'
+require_relative '../files_api_test_base' # Must be required first to establish load paths
+require_relative '../files_api_test_helper'
 require 'ostruct'
 
 #
