+++ shared/test/middleware/spy_pub_sub_api.rb
