+++ shared/test/middleware/test_animation_library_api.rb
@@ -1,4 +1,4 @@
-require_relative 'test_helper'
+require_relative '../test_helper'
 require 'animation_library_api'
 
 class AnimationLibraryTest < Minitest::Test
