+++ shared/test/middleware/test_animations.rb
