+++ shared/test/middleware/test_assets.rb
