+++ shared/test/middleware/test_files.rb
