+++ shared/test/middleware/test_libraries.rb
