+++ shared/test/middleware/test_redis_property_bag.rb
@@ -3,11 +3,11 @@
 # real instance on localhost by setting the USE_REAL_REDIS environment variable;
 # e.g. "USE_REAL_REDIS=true ruby test_redis_property_bag.rb".
 
-require_relative 'test_helper'
+require_relative '../test_helper'
 require 'fakeredis' unless use_real_redis?
 require 'timecop'
-require_relative 'fake_timecop'
-require_relative '../middleware/helpers/redis_property_bag'
+require_relative '../fake_timecop'
+require_relative '../../middleware/helpers/redis_property_bag'
 
 class RedisPropertyBagTest < Minitest::Test
   def setup
