+++ shared/test/middleware/test_redis_table.rb
@@ -5,10 +5,10 @@
 # Caution: This test is destructive to the Redis instance it runs against,
 # use caution when running against real redis.
 
-require_relative 'test_helper'
+require_relative '../test_helper'
 require 'fakeredis' unless use_real_redis?
 require 'timecop'
-require_relative 'fake_timecop'
+require_relative '../fake_timecop'
 require 'helpers/null_pub_sub_api'
 require 'helpers/redis_table'
 require_relative 'spy_pub_sub_api'
