+++ shared/test/middleware/test_tables.rb
@@ -1,5 +1,5 @@
 require 'mocha/mini_test'
-require_relative 'test_helper'
+require_relative '../test_helper'
 require 'channels_api'
 require 'tables_api'
